Refactor storage parser methods to StorageParser class
